pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mem0-analytics"
version = "0.1.0"
description = "Real-time observability and performance analytics for Mem0 AI memory systems"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Kaushal", email = "kaushal@example.com" }
]
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.3.0",
    "numpy>=2.3.0",
    "psutil>=7.0.0",
    "python-dotenv>=1.2.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.14.0",
    "SQLAlchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "posthog>=5.0.0",
    "openai>=2.0.0",
    "ollama>=0.6.0",
    "qdrant-client>=1.15.0",
    "chromadb>=1.3.0",
    "mem0ai>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/kaushal07wick/mem0-analytics"
Dashboard = "https://us.posthog.com/shared/0_gFtZ5fE8WhDNVXKlTHh2i4v31uSQ"

[tool.setuptools.packages.find]
where = ["src"]
